mlmodel/ml_service/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
pydantic
python-whois
requests
beautifulsoup4
pyOpenSSL 